Fix switched interfaces and implementations in service package
