+++ src/main/java/com/csemosip/bookingservice/service/Impl/AuthenticationServiceImpl.java
@@ -1,10 +1,44 @@
 package com.csemosip.bookingservice.service.Impl;
 
-
 import com.csemosip.bookingservice.dto.AuthDTO;
 import com.csemosip.bookingservice.dto.AuthenticationResponse;
+import com.csemosip.bookingservice.repository.UserRepository;
+import com.csemosip.bookingservice.service.AuthenticationService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.stereotype.Service;
 
-public interface AuthenticationServiceImpl {
-    AuthenticationResponse login(AuthDTO authDTO);
-    AuthenticationResponse logout();
+@Service
+@RequiredArgsConstructor
+public class AuthenticationServiceImpl implements AuthenticationService {
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    private JWTService jwtService;
+
+    @Autowired
+    private AuthenticationManager authenticationManager;
+
+    @Override
+    public AuthenticationResponse login(AuthDTO authDTO) {
+        AuthenticationResponse response = new AuthenticationResponse();
+        authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(
+                        authDTO.getUsername(),
+                        authDTO.getPassword()
+                )
+        );
+        var user = userRepository.findByUsername(authDTO.getUsername()).orElseThrow();
+        var token = jwtService.generateToken(user);
+        response.setToken(token);
+        return response;
+    }
+
+    @Override
+    public AuthenticationResponse logout() {
+        return null;
+    }
 }
