+++ src/main/java/com/csemosip/bookingservice/service/Impl/UserServiceImpl.java
@@ -1,14 +1,41 @@
 package com.csemosip.bookingservice.service.Impl;
 
 import com.csemosip.bookingservice.dto.UserDTO;
+import com.csemosip.bookingservice.exception.ResourceNotFoundException;
 import com.csemosip.bookingservice.model.User;
+import com.csemosip.bookingservice.repository.UserRepository;
+import com.csemosip.bookingservice.service.UserService;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
 
+@Service
+public class UserServiceImpl implements UserService {
 
-public interface UserServiceImpl {
-    User createUser(UserDTO userDTO) ;
+    @Autowired
+    private UserRepository userRepository;
 
-    User findUser(String id);
+    @Autowired
+    private ModelMapper modelMapper;
 
-    User findByUsername(String username);
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+    @Override
+    public User createUser(UserDTO userDTO) {
+        String encodedPassword = passwordEncoder.encode(userDTO.getPassword());
+        userDTO.setPassword(encodedPassword);
+        User user = modelMapper.map(userDTO, User.class);
+        return userRepository.save(user);
+    }
 
+    @Override
+    public User findUser(String id) {
+        return userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User Not Found"));
+    }
+
+    @Override
+    public User findByUsername(String username) {
+        return userRepository.findByUsername(username).orElseThrow(() -> new ResourceNotFoundException("User Not Found"));
+    }
 }
